chore(restructure): move translate -> openshorts/integrations/elevenlabs.py
Phase 1 step 2: relocate the ElevenLabs dubbing client. Only app.py
imports from translate. Adds a re-export shim at the old path. Tests
stay 62/62 green.
